introduce lenskit.state and document it
